+++ GlkView/GlkGraphicsWindow.m
@@ -156,12 +156,12 @@
 	lastSize = [self glkSize];
 }
 
-- (float) widthForFixedSize: (unsigned) size {
+- (CGFloat) widthForFixedSize: (unsigned) size {
 	// Graphics sizes are already in pixels
 	return size;
 }
 
-- (float) heightForFixedSize: (unsigned) size {
+- (CGFloat) heightForFixedSize: (unsigned) size {
 	// Graphics sizes are already in pixels
 	return size;
 }
